[lp-2189] : create personwithsignificantcontrol package and move classes into it
